Revert "Tweaks to run on Debian, clean up signed comparison."
This reverts commit c46e7625700351f07ac62d0ab87c6b8c135ab5cc.

+++ src/PuzzleBoard.cpp
@@ -1,5 +1,7 @@
 #include "PuzzleBoard.h"
 
+#include <bits/stdint-uintn.h>
+
 #include <cmath>
 #include <iostream>
 #include <stdexcept>
@@ -136,9 +138,9 @@ PuzzleBoard::PlayerMove(const std::string &pieceName, std::pair<uint32_t, uint32
   auto it2 = ComputerPieces.find(PuzzlePiece::PlayerOrientationToComputerOrientation[PuzzlePiece::PlayerPieceNameToOrientation[pieceName]]);
   auto currentComputerPosition = it2->second.GetPosition();
   auto newComputerPosition = currentComputerPosition;
-  if (static_cast<int32_t>(currentComputerPosition.second) + dx >= 0 || currentComputerPosition.second + dx < BoardSize)
+  if (currentComputerPosition.second + dx >= 0 || currentComputerPosition.second + dx < BoardSize)
     newComputerPosition.second = currentComputerPosition.second + dx;
-  if (static_cast<int32_t>(currentComputerPosition.first) - dy >= 0 || currentComputerPosition.first - dy < BoardSize)
+  if (currentComputerPosition.first - dy >= 0 || currentComputerPosition.first - dy < BoardSize)
     newComputerPosition.first = currentComputerPosition.first - dy;
 
   if (OccupiedPositions.find(newComputerPosition) == OccupiedPositions.end()) {
